Some cleanup work
* Switch to material icon for search
* Split SearchBox into smaller components
* Attempt to fix a complexity code smell for Code Climate
